resolved issues raised in PR review
